Added add_to_payload inside the LoraWAN handler
